requirements.txt
selenium==4.12.0
webdriver-manager==3.8.6
anyio==4.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
astroid==3.3.8
asttokens==2.4.1
async-lru==2.0.4
attrs==24.3.0
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.9
distlib==0.3.7
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
et_xmlfile==2.0.0
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.12.2
fonttools==4.46.0
fqdn==1.5.1
h11==0.14.0
idna==3.6
iniconfig==2.0.0
ipykernel==6.27.1
ipython==8.18.1
ipywidgets==8.1.1
isoduration==20.11.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.11.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==3.0.2
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.0
openpyxl==3.1.5
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
Pillow==10.1.0
pipenv==2023.9.8
platformdirs==3.10.0
pluggy==1.5.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
pylint==3.3.3
pyparsing==3.1.1
PySocks==1.7.1
pytest==8.3.4
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0
pyzmq==25.1.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.31.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.2
selenium==4.11.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
terminado==0.18.0
tinycss2==1.2.1
tomlkit==0.13.2
tornado==6.4
traitlets==5.14.0
trio==0.28.0
trio-websocket==0.11.1
types-python-dateutil==2.8.19.14
typing_extensions==4.12.2
tzdata==2023.3
uri-template==1.3.0
urllib3==2.1.0
virtualenv==20.24.2
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.9
wsproto==1.2.0
